% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_payments_by_DRG.R
\name{boxplot_payments_by_DRG}
\alias{boxplot_payments_by_DRG}
\title{Boxplot of Payments by DRG}
\usage{
boxplot_payments_by_DRG(df, payment_type)
}
\arguments{
\item{df}{a dataframe}

\item{payment_type}{a variable name, choose from 'Average.Medicare.Payments', 'Average.Total.Payments', or 'Average.Covered.Charges'}
}
\value{
A boxplot of the choosen payment by DRG code
}
\description{
This function produces a boxplot of payments by DRG code.
It can plot the average Medicare payments, the average total payment, or the average covered charges.
}
\examples{
boxplot_payments_by_DRG(DRG, "Average.Medicare.Payments")

}
