% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_calculate.R
\name{stat_calculate}
\alias{stat_calculate}
\title{Statistics Calculation}
\usage{
stat_calculate(df, varname, alternative)
}
\arguments{
\item{df}{a dataframe}

\item{varname}{a variable name of the payments}

\item{alternative}{Type of the statistics. Choose from 'mean', 'median', or 'standard deviation'}
}
\value{
A value of the choosen statistics
}
\description{
This function calculates statistics over all of the DRG codes for average Medicare payments.
It can calculate the mean, median, or standard deviation of the DRG codes.
}
\examples{
stat_calculate(DRG, "Average.Medicare.Payments", alternative = "mean")
}
